Clean up code styles
Add a CSS var `--code-bg` that is used in CodeSnippet and Markdown code blocks. (Because it's used in 2 places, a CSS var is used instead of just `--color-bg-2`.)

Also fix a bug where the CodeExcerpt style was not actually applied in the storybook because only the global stylesheet was being loaded.